Fixes getProfileDiff function (#599)
* Fixes getProfileDiff function

User data was visible without encryption. In this PR, Email and Phone number are somwwhat encypted.
Showing only first and last two characters of User's Phone number.
For emails, only first two characters and last four characters are visible.
Rest of the details of Email and Phone are hidden using Asterisk.

Approve and Timestamp fields are not allowed. They are not returning in getProfileDiff function
as a field in user data.

* Removed the firestore config file

* fixed conflicts

* Removed un-necessary change

* Removed un-necessary change